+++ examples/schema-field-update.rs
@@ -22,9 +22,7 @@
 //! $ cargo run --example schema-field-update task sg_hello name hello_world
 //! ```
 
-use serde_json::Value;
-use shotgun_rs::types::{CreateUpdateFieldProperty, UpdateFieldRequest};
-use shotgun_rs::Shotgun;
+use shotgun_rs::{Shotgun, TokenResponse};
 use std::env;
 
 #[tokio::main]
@@ -35,32 +33,25 @@ async fn main() -> shotgun_rs::Result<()> {
     let script_name = env::var("SG_SCRIPT_NAME").expect("SG_SCRIPT_NAME is required var.");
     let script_key = env::var("SG_SCRIPT_KEY").expect("SG_SCRIPT_KEY is required var.");
 
-    let entity_type: Option<String> = env::args().nth(1);
-    let field_name: Option<String> = env::args().nth(2);
-    let property_name: Option<String> = env::args().nth(3);
-    let value: Option<String> = env::args().nth(4);
+    let entity_type = env::args().nth(1).expect("entity type");
+    let field_name = env::args().nth(2).expect("field name");
+    let property_name = env::args().nth(3).expect("property name");
+    let value = env::args().nth(4).expect("property value");
 
-    println!("Attempting to update {:?} on {:?}", field_name, entity_type);
+    println!("Attempting to update {} on {}", field_name, entity_type);
 
     let sg = Shotgun::new(server, Some(&script_name), Some(&script_key)).expect("SG Client");
 
-    let token = {
-        let resp: Value = sg.authenticate_script().await?;
-        resp["access_token"].as_str().unwrap().to_string()
-    };
-
-    let properties = CreateUpdateFieldProperty {
-        property_name: property_name.unwrap(),
-        value: value.unwrap(),
-    };
-
-    let data = UpdateFieldRequest {
-        properties: vec![properties],
-        project_id: None,
-    };
+    let TokenResponse { access_token, .. } = sg.authenticate_script().await?;
 
     let resp = sg
-        .schema_field_update(&token, &entity_type.unwrap(), &field_name.unwrap(), &data)
+        .schema_field_update(
+            &access_token,
+            &entity_type,
+            &field_name,
+            vec![(property_name, value)],
+            None,
+        )
         .await?;
 
     println!("Data: {:?}", resp.data);