+++ src/types.rs
@@ -1,10 +1,14 @@
-use serde_json::Value;
-use std::collections::HashMap;
-
+pub use crate::schema::{
+    CreateFieldRequest, CreateUpdateFieldProperty, FieldDataType, SchemaEntitiesResponse,
+    SchemaEntityRecord, SchemaEntityResponse, SchemaFieldProperties, SchemaFieldRecord,
+    SchemaFieldResponse, SchemaFieldsResponse, SchemaResponseValue, UpdateFieldRequest,
+};
 pub use crate::summarize::{
     Grouping, GroupingDirection, GroupingType, SummarizeRequest, SummarizeResponse, SummaryData,
     SummaryField, SummaryFieldType, SummaryMap, SummaryOptions,
 };
+use serde_json::Value;
+use std::collections::HashMap;
 
 pub type Filters = Value; // FIXME: SGRS-32 need a more sophisticated representation for filters.
 
@@ -59,20 +63,6 @@ impl Default for ClientCredentialsRequest {
             client_secret: None,
         }
     }
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/#tocScreatefieldrequest>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CreateFieldRequest {
-    pub data_type: FieldDataType,
-    pub properties: Vec<CreateUpdateFieldProperty>,
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/#tocScreateupdatefieldproperty>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct CreateUpdateFieldProperty {
-    pub property_name: String,
-    pub value: String,
 }
 
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -135,49 +125,6 @@ pub struct ErrorObject {
     pub detail: Option<String>,
     pub source: Option<serde_json::Map<String, Value>>,
     pub meta: Option<serde_json::Map<String, Value>>,
-}
-
-/// How to perform the grouping for a given summary request.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum FieldDataType {
-    #[serde(rename = "checkbox")]
-    Checkbox,
-    #[serde(rename = "currency")]
-    Currency,
-    #[serde(rename = "date")]
-    Date,
-    #[serde(rename = "date_time")]
-    DateTime,
-    #[serde(rename = "duration")]
-    Duration,
-    #[serde(rename = "entity")]
-    Entity,
-    #[serde(rename = "float")]
-    Float,
-    #[serde(rename = "int")]
-    Int,
-    #[serde(rename = "list")]
-    List,
-    #[serde(rename = "multi_entity")]
-    MultiEntity,
-    #[serde(rename = "number")]
-    Number,
-    #[serde(rename = "percent")]
-    Percent,
-    #[serde(rename = "status_list")]
-    StatusList,
-    #[serde(rename = "text")]
-    Text,
-    #[serde(rename = "timecode")]
-    Timecode,
-    #[serde(rename = "footage")]
-    Footage,
-    #[serde(rename = "url")]
-    URL,
-    #[serde(rename = "uuid")]
-    UUID,
-    #[serde(rename = "calculated")]
-    Calculated,
 }
 
 /// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSfieldhashresponse>
@@ -460,58 +407,6 @@ pub enum ReturnOnly {
     Retired,
 }
 
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#schemaschemaentityrecord>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SchemaEntityRecord {
-    pub name: Option<SchemaResponseValue>,
-    pub visible: Option<SchemaResponseValue>,
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSschemaentityresponse>
-pub type SchemaEntityResponse = SingleResourceResponse<SchemaEntityRecord, SelfLink>;
-
-/// <https://developer.shotgunsoftware.com/rest-api/#tocSschemaentitiesresponse>
-pub type SchemaEntitiesResponse = ResourceMapResponse<SchemaEntityRecord, SelfLink>;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SchemaFieldProperties {
-    pub default_value: Option<SchemaResponseValue>,
-    pub regex_validation: Option<SchemaResponseValue>,
-    pub regex_validation_enabled: Option<SchemaResponseValue>,
-    pub summary_default: Option<SchemaResponseValue>,
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSschemafieldrecord>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SchemaFieldRecord {
-    pub custom_metadata: Option<SchemaResponseValue>,
-    pub data_type: Option<SchemaResponseValue>,
-    pub description: Option<SchemaResponseValue>,
-    pub editable: Option<SchemaResponseValue>,
-    pub entity_type: Option<SchemaResponseValue>,
-    pub mandatory: Option<SchemaResponseValue>,
-    pub name: Option<SchemaResponseValue>,
-    pub properties: Option<SchemaFieldProperties>,
-    pub ui_value_displayable: Option<SchemaResponseValue>,
-    pub unique: Option<SchemaResponseValue>,
-    pub visible: Option<SchemaResponseValue>,
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSschemafieldresponse>
-pub type SchemaFieldResponse = SingleResourceResponse<SchemaFieldRecord, SelfLink>;
-
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSschemafieldsresponse>
-pub type SchemaFieldsResponse =
-    SingleResourceResponse<HashMap<String, SchemaFieldRecord>, SelfLink>;
-
-/// <https://developer.shotgunsoftware.com/rest-api/?shell#schemaschemaresponsevalue>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct SchemaResponseValue {
-    /// Can be a string or a boolean
-    pub value: Option<Value>,
-    pub editable: Option<bool>,
-}
-
 /// <https://developer.shotgunsoftware.com/rest-api/?shell#tocSsearchrequest>
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct SearchRequest {
@@ -559,14 +454,6 @@ pub struct TextSearchRequest {
     pub text: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub sort: Option<String>,
-}
-
-/// <https://developer.shotgunsoftware.com/rest-api/#tocSupdatefieldrequest>
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct UpdateFieldRequest {
-    pub properties: Vec<CreateUpdateFieldProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub project_id: Option<i32>,
 }
 
 /// <https://developer.shotgunsoftware.com/rest-api/#tocSupdateworkdayrulesrequest>